src/edge/filter/pyproject.toml
[project]
name = "filter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=2.1.0",
]
